backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
pymysql>=1.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]
bcrypt==4.0.1
python-multipart>=0.0.6
cryptography
pydantic>=2.0.0
pydantic[email]
pydantic-settings>=2.0.0


pandas
numpy
numpy_financial